2020.02.04 하던거

+++ src/stores/epics/auth.ts
@@ -1,45 +1,42 @@
 import { combineEpics } from 'redux-observable';
-import { EMPTY, of } from 'rxjs';
+import { of } from 'rxjs';
 import { catchError, filter, map, switchMap, tap } from 'rxjs/operators';
 import { isActionOf } from 'typesafe-actions';
 import { authTokenStorageKey } from '../../constants';
 import { createStorage } from '../../utils';
-import { authorizeWithTokenActions, authorizeWithTokenWhichFromStorageAction } from '../actions';
+import { authorizeAction, authorizeWithTokenWhichFromStorageAction } from '../actions';
 import { Epic } from '../types';
 
 const storage = createStorage();
 
-const findTokenFromStorageAndAuthorizeEpic: Epic = action$ =>
+const findTokenFromStorageAndAuthorizeEpic: Epic = (action$, _, { api }) =>
   action$.pipe(
     filter(isActionOf(authorizeWithTokenWhichFromStorageAction)),
     switchMap(() => {
       const token = storage.get(authTokenStorageKey);
 
-      return token !== null ? of(authorizeWithTokenActions.request({ token })) : EMPTY;
-    }),
-  );
+      if (token === null) {
+        return of(authorizeAction({ authorized: false }));
+      }
 
-const authorizeEpic: Epic = (action$, _, { api }) =>
-  action$.pipe(
-    filter(isActionOf(authorizeWithTokenActions.request)),
-    switchMap(({ payload: { token } }) =>
-      api.authorize(token).pipe(
+      return api.authorize(token).pipe(
         tap(() => {
           storage.set(authTokenStorageKey, token);
         }),
         map(user =>
-          authorizeWithTokenActions.success({
+          authorizeAction({
+            authorized: true,
             user,
             token,
           }),
         ),
-        catchError(error => {
+        catchError(() => {
           storage.delete(authTokenStorageKey);
 
-          return of(authorizeWithTokenActions.failure({ error }));
+          return of(authorizeAction({ authorized: false }));
         }),
-      ),
-    ),
+      );
+    }),
   );
 
-export const authEpic = combineEpics(findTokenFromStorageAndAuthorizeEpic, authorizeEpic);
+export const authEpic = combineEpics(findTokenFromStorageAndAuthorizeEpic);
